[refactor] Reorg components into subfolders
