+++ client/src/components/grid/Laser.jsx
